extract user menu as separated component file
